Extracting OltPortStatus from OltFlowService and provided equality checks
